Increase size of RISCV x registers array by 1
In order to avoid hard-to-catch off-by-one errors in accessing the x
register of the guest CPU state, the size of the x field will be
increased to 32. This also removes the need to offset the register
access in the helper functions. The check for 0 is left in since we
don't want to read or write it directly. We only need to check for raw
accesses to 0 now, which should be easier to spot.

Test: mm and tree-hugger
Bug: 279808127
Change-Id: I7cf2bc082bdcae68b34788bd9377ae74f9acca39