base: Fix concurrency bug in MmapPool
LockFreeStack must not be used with nodes that are reused
or deallocated after popping out of the stack. This is because
threads simultaneously accessing LockFreeStack may still hold
the pointer to the popped node and try to dereference it (node->next).

So instead of using blocks themselves as LockFreeStack nodes, we
preallocate persitent array of auxilary nodes, which we associate with
blocks and move to and from the stack.

Also instead of imprecise asynchronious pool size calculation, we
introduce another LockFreeStack of nodes available for allocation
that accurately limits the total amount of memory that can be allocated
by the pool.

Test: Added stress test
Bug: 287721740
Change-Id: I1138f393402f19930e35a669f91876c7eb2f0cf7