tests: Disable NDK program tests in berberis_all
NDK program tests are currently broken in post-submit because they are
built for riscv64 with the stock configuration, which includes bit
manipulation extensions.  Berberis currently does not support all of
these instructions, so tests fail unless the extensions are explicitly
disabled.  To circumvent this, the program tests are temporarily
excluded from berberis_all so they do not run in post-submit.

Test: berberis_all does not run berberis_ndk_program_tests
Bug: 295103992
Change-Id: I2e1b4486eb7546a8c9b76c85dff4a5139fd6555b