code_gen_lib: NaN-box 32-bit floats in FP regs
The LP64D calling convention specifies that floating-point arguments
shorter than FLEN that are passed in floating-point registers must be
NaN-boxed by 1-extentension.  In practice this means that 32-bit floats
in FP registers must have 0xffffffff in the high bits.  However, XMM
does not use NaN boxing.  We must account for this during wrapper
generation by filling the high bits when copying the XMM arguments to
the FP registers and only copying the lower 32 bits of the FP result
register to the XMM result register.

Test: berberis_all
Bug: 282063730
Change-Id: Iddd9e4c55cd9ff63b5d9e8c65b06b8849b6ad4e4