runtime: Fix SIGSEGV in executable memory check
TranslateRegion was dereferencing memory at the current PC without
knowing whether it was in an executable region.  This caused a SIGSEGV
in the translator when jumping to an unmapped address.  The reason for
the dereference was to determine the size of the instruction (i.e. 2 or
4 bytes).  Instead, check whether the first two bytes at PC are
executable, then get the instruction size and check whether the
remainder is executable.

Test: berberis_run_host_tests
Bug: 300488860
Change-Id: I16977638d52a57ec0744d60fe25633066e581029