guest_abi: Fix extension for integer arguments
RISC-V integer arguments narrower than 64 bits must be widened to 32
bits according to their signedness (zero-extended for unsigned,
sign-extended for signed), then sign-extended to 64 bits.  This was not
implemented in the RISC-V guest ABI.  To fix this, GuestArgumentInfo now
selects a GuestArgument specialization that properly boxes and unboxes
values within a 64-bit storage, which represents either a register in
CPUState or an argument in GuestArgumentBuffer.

Test: berberis_run_host_tests
Bug: 295536668
Change-Id: Ifafe72b947f129b3a367eed22b873bdb8650968e