Add BitMaskToSimdMask function.
On x86-64 platform bitmasks are only supported in AVX10 which we
couldn't assume to be available.

BMI2 instructions can also be used to effectively generate masks,
but they are only usable on AMD CPUs starting from Zen3.

Thus we need SSE2-based baseline which can be implemented with lookup
tables.

Test: m berberis_all

Change-Id: I56049d926aa14e70672db92476474b03d0cbdf6a