Extend AlignUp/AlignDown function to cover safe ints.
Most operations couldn't be performed in just, e.g., 8bit integer: you
have to know whether you are dealing wuth signed integer or unsigned one
and/or you have to decide whether you need to handle overflow in saturation
way or wrapping way.

But alignment operations are different: they are not overflowing and not
ambigious.

Bug: 318164404

Test: m berberis_all

Change-Id: I9c0efe1c1050cf36d0eecbe73d44143f25807f9b