Rename helper defines to separate inputs and outputs.
Fused multiply-add intrunsic make it ambigious: what does WVV suffix mean?
Technically they don't class because vwmaccsu.vx needs VVW suffix which is
different from WVV and WWV suffixes, but mix of inputs and outputs make it
hard to to understand what suffix is supposed to say.

Width of output (w or n) is specified in the RISC-V instruction after first
V letter and it makes ssense to use the same scheme in our intrinsics.

Test: berberis_all

Change-Id: I7c2d810e2f2cecc3ddbc641da50bb265cb2e75a9