Convert `berberis_all` to Android.bp
Convert the part in the top half of this Android.mk that adds
dependencies to `berberis_all` to Android.bp.

Move the part that executes `berberis_host_tests_result` and
`berberis_run_host_tests` in the bottom half to `build/core/tasks`,
as Soong might currently have difficulty supporting this part.

Bug: 321643544
Test: 1. lunch sdk_phone64_x86_64_riscv64-trunk_staging-eng
      2. m berberis_all
      3. m berberis_host_tests_result
      4. m berberis_run_host_tests
Change-Id: I413cb52e7b6bd938ef4eba975431dddc8ab490a4