decoder: rename Unimplemented to Undefined
Originally Unimlemented was supposed to be a hint about
instructions we still need to implement. But over time it
started to be used for instructions undefined in riscv64 ISA.
There are still some unimplemented instructions but vast majority
of Unimplemented uses actually mean Undefined. So let's rename
everything to Undefined to avoid misleading hinting about
unimplemented opcodes. Instructions that are actually not yet
implemented are tracked independently anyway.

Test: tree-hugger
Bug: 232598137
Change-Id: Idc6848b11e4cb37c1d48dbe301a3fd59a5a19ac8