inline_asm_tests: first riscv test
These tests can be run both under berberis and on
real devices. This way the test expectations themselves can be
verified.

To bootstrap we are porting vredsum tests from interpreter. The
idea is to keep the expectations and test structure, but
replace CPUState accesses with embedded assembly.

Note that vredsum is known to be incorrectly implemented (swapped
arguments), but we intentionally keep the expectations as is
to make it pass on berberis.

Masked tests are not supported yet.

Also fix a one-liner bug in interpreter triggered by this test.

Test: This is the test
Bug: 301577077

Change-Id: I81db036d4b7cabae6d08235f77f1126d81f8f584