Translate /proc/cpuinfo for executables
In zygote this is not needed because native bridge is mounting
/proc/cpuinfo to point to the emulated version. But for executables
this does not happen and they end up reading host cpuinfo. Note that
current selinux policies prevent us from just mounting /proc/cpuinfo
for executables.

This change translates /proc/cpuinfo to the guest version
when emulating openat syscall for the guest.

Bug: http://b/331209929
Test: run CtsSimpleperfTestCases
Change-Id: I44cf3a16f1c9d6dc5a5ea12544432a182c8ef180