Add Shift32/Shift64-immediates to RISC V assembler.
In RISC V manual shifts are described as using I-format with complex
restrictions for which immediates are accepted and allowed (with parts
of what manual classifies as “immediate” used to determine the actual
instruction used and rules which differ between RV32 and RV64!).

Instead of doing special handling for the instructions in python scripts
we just reclassify these parts of immediate as “opcode” and reclassify
these instructions as “Shift32-type” and “Shift64-type”.

This also means that the same instructions for RV32 and RV64 would have
different types, but since we don't have a goal to make RV32 a strict subset
of RV64 that's acceptable.

In addition we provide aliases in RV32 and RV64 assemblers to make sure
users of assembler may still use ShiftImmediate and make_shift_immediate
for native width without thinking about details of implementation.

Bug: http://b/232598137

Test: m berberis_all

Change-Id: I2e48d199e3d8a0b74f81b267723e024010251251