Minor style cleanups.
Currently we don't have code that uses GetPhysicalIndex with RISC-V, but
we have similar code in x86 case thus it probably makes sense to expose
it.

And in Text Assembler we only use default constructor in a very few
cases thus it's probably better to make the fact that we are not getting
usable register there explicit.

Test: m berberis_all

Change-Id: I3b87cd97148ef2736581f3d1da521dcb4fb81ddb