Add kImpossible{Type,Value}Const to dependent_false.h
Every const in C++ have to have a value assigned if it's not extern.

Usually we define consts that are not always present as extern and then
attempt to use them is an error, but a link-time error.

Compile-time error is preferable to link-time error, though and the
usual trick, similar to kDependentFalse makes it possible to use it.

Also add one missing const to macro_assembler_constants_pool.h

Test: m berneris_all

Change-Id: I84d7205deb60bdbed472c231b689e7e7fc456b18