Use tuples instead of arrays in tests with SIMD128Register.
After https://r.android.com/3256885 it's possible to use SIMD128Register in guest code and thus in
inline_asm_tests_riscv64, but SIMD128 used in main_riscv64.cc and SIMD128Register are incompatible.

The main difference is the use of arrays or tuples for constants.

While it sounds as if arrays are more natural for sequence of 2/4/8/16 numbers in reality tuples
are safer because in C++ arrays are allowed to have less initializers than there are elements,
but tuples have to have exactly as many arguments as there are elements.

And because of that tuples can also resolve ambiguities: it's not possible to do something like
SIMD128Register x = {{0, 1, 2, 3}} - because it can be interpreted as four int32, eighth int16 or
even sixteen int8 values. Compiler couldn't choose and code is rejected.

With tuples there are no ambiguity: SIMD128Register x = {{0, 1, 2, 3}} means it's int32x4 version.

Test: m berberis_all

Change-Id: Icd35edbd8f63abcb6e99ce090b8db2e523a25ad2