Add separate flags to control ir and addrs printing
Often we don't need to print full IRs, and just the translated
addrs is enough. So we add flags for both to use in the desired
combination. It's unlikely that someone needs PrintIRs=true and
PrintTranslatedAddrs=false, but we still keep two flags for simplicity.

verbose-translation behaves as before.

Flag: EXEMPT NDK
Test: trigger prints in different combinations and check the resutl
Bug: 395189011
Change-Id: If4902aca8ce307808bf3c2938e31ac34d1f928cf